+++ engine/server/gc.go
@@ -1,0 +1,156 @@
+package server
+
+import (
+	"context"
+	"time"
+
+	controlapi "github.com/moby/buildkit/api/services/control"
+	"github.com/moby/buildkit/client"
+	bkclient "github.com/moby/buildkit/client"
+	"github.com/moby/buildkit/cmd/buildkitd/config"
+	"github.com/moby/buildkit/util/bklog"
+	"github.com/moby/buildkit/util/imageutil"
+	"golang.org/x/sync/errgroup"
+)
+
+func (e *Engine) DiskUsage(ctx context.Context, r *controlapi.DiskUsageRequest) (*controlapi.DiskUsageResponse, error) {
+	resp := &controlapi.DiskUsageResponse{}
+	du, err := e.baseWorker.DiskUsage(ctx, bkclient.DiskUsageInfo{
+		Filter: r.Filter,
+	})
+	if err != nil {
+		return nil, err
+	}
+	for _, r := range du {
+		resp.Record = append(resp.Record, &controlapi.UsageRecord{
+			ID:          r.ID,
+			Mutable:     r.Mutable,
+			InUse:       r.InUse,
+			Size_:       r.Size,
+			Parents:     r.Parents,
+			UsageCount:  int64(r.UsageCount),
+			Description: r.Description,
+			CreatedAt:   r.CreatedAt,
+			LastUsedAt:  r.LastUsedAt,
+			RecordType:  string(r.RecordType),
+			Shared:      r.Shared,
+		})
+	}
+	return resp, nil
+}
+
+func (e *Engine) Prune(req *controlapi.PruneRequest, stream controlapi.Control_PruneServer) error {
+	eg, ctx := errgroup.WithContext(stream.Context())
+
+	e.serverMu.RLock()
+	cancelLeases := len(e.servers) == 0
+	e.serverMu.RUnlock()
+	if cancelLeases {
+		imageutil.CancelCacheLeases()
+	}
+
+	didPrune := false
+	defer func() {
+		if didPrune {
+			if e, ok := e.SolverCache.(interface {
+				ReleaseUnreferenced(context.Context) error
+			}); ok {
+				if err := e.ReleaseUnreferenced(ctx); err != nil {
+					bklog.G(ctx).Errorf("failed to release cache metadata: %+v", err)
+				}
+			}
+		}
+	}()
+
+	ch := make(chan bkclient.UsageInfo, 32)
+
+	eg.Go(func() error {
+		defer close(ch)
+		return e.baseWorker.Prune(ctx, ch, bkclient.PruneInfo{
+			Filter:       req.Filter,
+			All:          req.All,
+			KeepDuration: time.Duration(req.KeepDuration),
+			KeepBytes:    req.KeepBytes,
+		})
+	})
+
+	eg.Go(func() error {
+		defer func() {
+			// drain channel on error
+			for range ch {
+			}
+		}()
+		for r := range ch {
+			didPrune = true
+			if err := stream.Send(&controlapi.UsageRecord{
+				ID:          r.ID,
+				Mutable:     r.Mutable,
+				InUse:       r.InUse,
+				Size_:       r.Size,
+				Parents:     r.Parents,
+				UsageCount:  int64(r.UsageCount),
+				Description: r.Description,
+				CreatedAt:   r.CreatedAt,
+				LastUsedAt:  r.LastUsedAt,
+				RecordType:  string(r.RecordType),
+				Shared:      r.Shared,
+			}); err != nil {
+				return err
+			}
+		}
+		return nil
+	})
+
+	return eg.Wait()
+}
+
+func (e *Engine) gc() {
+	e.gcmu.Lock()
+	defer e.gcmu.Unlock()
+
+	ch := make(chan bkclient.UsageInfo)
+	eg, ctx := errgroup.WithContext(context.TODO())
+
+	var size int64
+	eg.Go(func() error {
+		for ui := range ch {
+			size += ui.Size
+		}
+		return nil
+	})
+
+	eg.Go(func() error {
+		defer close(ch)
+		if policy := e.baseWorker.GCPolicy(); len(policy) > 0 {
+			return e.baseWorker.Prune(ctx, ch, policy...)
+		}
+		return nil
+	})
+
+	err := eg.Wait()
+	if err != nil {
+		bklog.G(ctx).Errorf("gc error: %+v", err)
+	}
+	if size > 0 {
+		bklog.G(ctx).Debugf("gc cleaned up %d bytes", size)
+	}
+}
+
+func getGCPolicy(cfg config.GCConfig, root string) []client.PruneInfo {
+	if cfg.GC != nil && !*cfg.GC {
+		return nil
+	}
+	if len(cfg.GCPolicy) == 0 {
+		cfg.GCPolicy = config.DefaultGCPolicy(cfg.GCKeepStorage)
+	}
+	out := make([]client.PruneInfo, 0, len(cfg.GCPolicy))
+	for _, rule := range cfg.GCPolicy {
+		out = append(out, client.PruneInfo{
+			Filter:       rule.Filters,
+			All:          rule.All,
+			KeepBytes:    rule.KeepBytes.AsBytes(root),
+			KeepDuration: rule.KeepDuration.Duration,
+		})
+	}
+	return out
+}
