+++ engine/server/platform.go
@@ -1,0 +1,26 @@
+package server
+
+import (
+	"github.com/containerd/containerd/platforms"
+	ocispecs "github.com/opencontainers/image-spec/specs-go/v1"
+)
+
+func parsePlatforms(platformsStr []string) ([]ocispecs.Platform, error) {
+	out := make([]ocispecs.Platform, 0, len(platformsStr))
+	for _, s := range platformsStr {
+		p, err := platforms.Parse(s)
+		if err != nil {
+			return nil, err
+		}
+		out = append(out, platforms.Normalize(p))
+	}
+	return out, nil
+}
+
+func FormatPlatforms(p []ocispecs.Platform) []string {
+	str := make([]string, 0, len(p))
+	for _, pp := range p {
+		str = append(str, platforms.Format(platforms.Normalize(pp)))
+	}
+	return str
+}
